[CAYC] Move XAML generator tests to match product folder structure
Fixed up namespaces
No product or test code changes
