+++ src/Education.UnitTests/XamlGenerator/RichRuleHelpXamlBuilderTests.cs
@@ -12,7 +12,7 @@ using SonarLint.VisualStudio.Education.XamlGenerator;
 using SonarLint.VisualStudio.Rules;
 using SonarLint.VisualStudio.TestInfrastructure;
 
-namespace SonarLint.VisualStudio.Education.UnitTests
+namespace SonarLint.VisualStudio.Education.UnitTests.XamlGenerator
 {
     [TestClass]
     public class RichRuleHelpXamlBuilderTests
