+++ src/Education.UnitTests/XamlGenerator/RuleHelpXamlBuilderTests.cs
@@ -25,7 +25,7 @@ using SonarLint.VisualStudio.Education.XamlGenerator;
 using SonarLint.VisualStudio.Rules;
 using SonarLint.VisualStudio.TestInfrastructure;
 
-namespace SonarLint.VisualStudio.Education.UnitTests;
+namespace SonarLint.VisualStudio.Education.UnitTests.XamlGenerator;
 
 [TestClass]
 public class RuleHelpXamlBuilderTests
