+++ src/Education.UnitTests/XamlGenerator/SimpleRuleHelpXamlBuilderTests.cs
@@ -32,7 +32,7 @@ using SonarLint.VisualStudio.Education.XamlGenerator;
 using SonarLint.VisualStudio.Rules;
 using SonarLint.VisualStudio.TestInfrastructure;
 
-namespace SonarLint.VisualStudio.Education.UnitTests
+namespace SonarLint.VisualStudio.Education.UnitTests.XamlGenerator
 {
     [TestClass]
     public class SimpleRuleHelpXamlBuilderTests
