+++ src/Education.UnitTests/XamlGenerator/XamlWriterFactoryTests.cs
@@ -24,7 +24,7 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 using SonarLint.VisualStudio.Education.XamlGenerator;
 using SonarLint.VisualStudio.TestInfrastructure;
 
-namespace SonarLint.VisualStudio.Education.UnitTests;
+namespace SonarLint.VisualStudio.Education.UnitTests.XamlGenerator;
 
 [TestClass]
 public class XamlWriterFactoryTests
